clean + fast star system

+++ MobilSemProjekt.MVVM/ViewModel/IRatingRestService.cs
@@ -7,7 +7,7 @@ namespace MobilSemProjekt.MVVM.ViewModel
     public interface IRatingRestService
     {
         Task Create(Rating rating);
-        Task<int> GetAverageRating(Location location);
+        Task<double> GetAverageRating(Location location);
         
         Task Update(Rating rating, int ratingId);
     }
